# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: TreePage API
  version: 0.1.0
  description: Documentation platform REST API
servers:
  - url: http://localhost:8081
    description: Auth service (dev)
  - url: http://localhost:8082
    description: Server API (dev)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Space:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        is_public:
          type: boolean
...
paths:
...
  /api/spaces:
    get:
      tags: [Spaces]
      summary: List accessible spaces
      security:
        - bearerAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Space"
    post:
      tags: [Spaces]
      summary: Create space
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slug, name]
              properties:
                slug:
                  type: string
                name:
                  type: string
                description:
                  type: string
                is_public:
                  type: boolean
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Space"
  /api/spaces/{slug}:
    get:
      tags: [Spaces]
      summary: Get space by slug
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Space"
...
  /api/admin/system-settings:
    get:
      tags: [Admin]
      summary: Get platform system settings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Auth, git, platform settings, and ui_theme
    put:
      tags: [Admin]
      summary: Update system settings (super_admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Updated settings
  /api/admin/system-settings/ui-theme:
    put:
      tags: [Admin]
      summary: Update global UI theme (super_admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ui_theme]
              properties:
                ui_theme:
                  type: string
                  enum: [coral_night, fox_white, light, dark]
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  ui_theme:
                    type: string
  /api/admin/repositories:
    get:
      tags: [Admin]
      summary: List all repositories with space info
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Repository list
    post:
      tags: [Admin]
      summary: Create repository and bind to space
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Created
  /api/admin/repositories/{id}:
    get:
      tags: [Admin]
      summary: Get repository with latest sync job
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Repository detail
    put:
      tags: [Admin]
      summary: Update repository
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Updated
    delete:
      tags: [Admin]
      summary: Delete repository (super_admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Deleted
  /api/admin/spaces/{id}/bind-repo:
    post:
      tags: [Admin]
      summary: Assign repository to a space
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [repository_id]
              properties:
                repository_id:
                  type: string
                  format: uuid
      responses:
        "200":
          description: Bound
  /api/admin/sync/{repoId}:
    post:
      tags: [Admin]
      summary: Trigger manual repository sync via sync service
      security:
        - bearerAuth: []
      parameters:
        - name: repoId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Sync triggered
  /api/admin/oidc-providers:
    get:
      tags: [Admin]
      summary: List OIDC providers (super_admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Provider list
    post:
      tags: [Admin]
      summary: Create OIDC provider (super_admin)
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Created
  /api/admin/users:
    get:
      tags: [Admin]
      summary: List users (super_admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User list
  /api/admin/spaces:
    get:
      tags: [Admin]
      summary: List all spaces (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Space list
  /api/admin/spaces/{id}:
    patch:
      tags: [Admin]
      summary: Update space (visibility, name)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                is_public:
                  type: boolean
      responses:
        "200":
          description: Updated space
...
  /api/admin/rag/status:
    get:
      tags: [Admin]
      summary: RAG indexing worker status
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Worker phase and progress
  /api/admin/rag/reindex:
    post:
      tags: [Admin]
      summary: Trigger full RAG reindex (super_admin)
      security:
        - bearerAuth: []
      responses:
        "202":
          description: Reindex started